chore(web): translation management handle undefined workflow variables object
