+++ apps/web/src/api/hooks/index.ts
@@ -2,3 +2,4 @@ export * from './notification-templates';
 export * from './useInAppActivated';
 export * from './useDeleteIntegration';
 export * from './useWebhookSupportStatus';
+export * from './notification-templates';
