+++ apps/web/src/api/hooks/notification-templates/index.ts
@@ -2,3 +2,4 @@ export { useTemplateFetcher } from './useTemplateFetcher';
 export { useUpdateTemplate } from './useUpdateTemplate';
 export * from './useFetchBlueprints';
 export * from './useCreateTemplateFromBlueprint';
+export * from './useWorkflowVariables';
